Add wrapper script to use during development
The 'dev/shmark.sh' wrapper script just sources the main shmark.sh
functions file and runs the main shmark() function. It's to be used
during development so that the functions file won't have to be sourced
repeatedly into the shell environment after every change. And also to
prevent the shell from exiting if an error occurs while running the
wrapper script with extra options such as 'errexit' and 'nounset'.