Fix help and cd action re list position shortcuts
The help documentation showed the hyphen-number syntax (-#) for all
actions that accept a list position argument. That syntax is only for
use as a standalone option, not an action arg. The actions should just
accept a number without a hyphen. Changed the 'cd' command to reflect
the update. The 'delete' action was already correct even though its
documentation was wrong.