Fix: Make sure -f flag only has temporary effect
When using the -f flag to specify a different bookmarks file than the
one from the SHMARK_FILE environment variable, that environment variable
was not being reset to its original value. This bug slipped through
testing because the -f flag is mostly just for testing, and running the
tests doesn't actually change the current shell environment like calling
a shell function directly does.